Add `FunnelChart` to `@dub/ui/charts`
