+++ packages/ui/src/charts/index.ts
@@ -1,6 +1,7 @@
 export * from "./areas";
 export * from "./bars";
 export * from "./chart-context";
+export * from "./funnel-chart";
 export * from "./time-series-chart";
 export * from "./x-axis";
 export * from "./y-axis";
